// packages/core/src/llm-core/agent/types.d.ts
export interface ChatCompletionMessageToolCall {
    /**
     * The ID of the tool call.
     */
    id: string
    /**
     * The function that the model called.
     */
    function: ChatCompletionMessageToolCall.Function
    /**
     * The type of the tool. Currently, only `function` is supported.
     */
    type: 'function'
}
/**
 * The name and arguments of a function that should be called, as generated by the
 * model.
 */
export interface ChatCompletionMessageFunctionCall {
    /**
     * The arguments to call the function with, as generated by the model in JSON
     * format. Note that the model does not always generate valid JSON, and may
     * hallucinate parameters not defined by your function schema. Validate the
     * arguments in your code before calling your function.
     */
    arguments: string
    /**
     * The name of the function to call.
     */
    name: string
}
export declare namespace ChatCompletionMessageToolCall {
    /**
     * The function that the model called.
     */
    interface Function {
        /**
         * The arguments to call the function with, as generated by the model in JSON
         * format. Note that the model does not always generate valid JSON, and may
         * hallucinate parameters not defined by your function schema. Validate the
         * arguments in your code before calling your function.
         */
        arguments: string
        /**
         * The name of the function to call.
         */
        name: string
    }
}
// ...
export type AgentAction = {
    tool: string
    toolInput: string | Record<string, any>
    toolCallId?: string
    log: string
}
export type AgentFinish = {
    returnValues: Record<string, any>
    log: string
}